# bicycle-api/docs/swagger.yaml
basePath: /v1
definitions:
  models.BicycleData:
    properties:
      revolutions:
        example: 10
        type: integer
      rpm:
        example: 200
        type: integer
      time:
        example: "2021-10-26T22:17:36Z"
        type: string
    type: object
  models.ConfigData:
    properties:
      radius_in_cm:
        example: 30
        type: number
      target_km:
        example: 150
        type: integer
    type: object
  models.ErrorResponse:
    properties:
      error_code:
        example: UnprocessableEntity
        type: string
      message:
        example: Malformed body
        type: string
      status_code:
        example: 422
        type: integer
    type: object
info:
  contact: {}
  title: bicycle API
  version: 0.1.0
paths:
  /bicycle:
    get:
      description: Gets all the bicycle data from database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BicycleData'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Gets all bicycle data from database
      tags:
      - BicyleData
  /bicycle/config:
    get:
      description: Gets the current config from database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ConfigData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Gets the current config from database
      tags:
      - ConfigData
  /bicycle/speed:
    get:
      description: Computes the current bicycle speed from the latest data in firebase
      responses:
        "200":
          description: OK
          schema:
            type: int
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Computes the current bicycle speed from database
      tags:
      - BicyleSpeed
swagger: "2.0"